+++ projects/np-ui-lib/src/lib/np-textarea/np-textarea.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { NpTextareaComponent } from './np-textarea.component';
+import { NpUtilityModule } from '../np-utility/np-utility.module';
+
+@NgModule({
+  declarations: [NpTextareaComponent],
+  imports: [
+    CommonModule,
+    NpUtilityModule
+  ],
+  exports: [NpTextareaComponent]
+})
+export class NpTextareaModule { }